// Revert normalization
git-svn-id: http://dev.prestashop.com/svn/v1/branches/1.5.x@8019 b9a71923-0436-4b27-9f14-aed3839534dd

+++ modules/blocklanguages/blocklanguages.php
@@ -62,7 +62,7 @@ class BlockLanguages extends Module
 			return;
 		$link = new Link();
 			
-		if (Configuration::get('PS_REWRITING_SETTINGS'))
+		if((int)Configuration::get('PS_REWRITING_SETTINGS'))
 		{
 			$default_rewrite = array();
 			$phpSelf = isset($_SERVER['PHP_SELF']) ? substr($_SERVER['PHP_SELF'], strlen(__PS_BASE_URI__)) : '';
